Change Model from SourceQuery to SourceQuerier
The domain model SourceQuery contains information to instantiate an
object to query a specific source. When performing this action the
naming of both the domain model and the resulting object leads to
confusion.

As the domain model describes a query, but is not actually one it
self, it shall be called SourceQuerier.